2 lepton channel control regions (#143)
* Initial commit, improve env.sh

As a small quality of life change, the env.sh script no longer requires
the user to change into its parent directory to source it.

* adding --site=DESY capability

* adding desy specific files for the tutorial

* enable single file for demo

* desy samples commented in many config

* fix: no site-dependent directory creation

* Uncomment TTbar sample for testing

* Style consistency for VHbbAnalysis.h

Removed spurious whitespace, added indentations, modified spacings of
parameters, added include guard.

Also removed the "using" directive for the "std" namespace. In many
places, the std prefix was already used, so may as well follow best
practice. This change is propagated to VHbbAnalysis.cc and compiles and
runs.

* Style and bugfix for VHbbAnalysis::Preselection()

Modified the code style in the Preselection() method for consistency.

Removed the boolean local variable "sel". This was set to true at the
beginning of the method and returned if the events passed the
preselection. It was never modified or used within the method after its
initialization. The method now directly returns true if the events pass
the preselection.

Removed blocks of code that were commented out. These lines have been
committed to a version control system and can be revisited at any
moment.

Fixed the 0-lepton trigger selection logic. Previously, one of the
conditions checked for a trigger bit to be not equal to -1. Now the
condition correctly checks that the trigger bit is not equal to 1.

The plugin compiles and runs.

* First pass at VHbbAnalysis::Analyze()

Modified the following up to line 264:

* Apply consistent code style.
* Removed duplicate CMVAv2 to CSV reassignment.
* Fixed 0-lepton trigger selection.

The plugin compiles and runs.

* Second pass at VHbbAnalysis::Analyze()

Modified the following up to line 464:

* Apply consistent code style.
* Removed a chunk of commented code which already resides in version
  control.

The plugin compiles and runs.

* Third pass at VHbbAnalysis::Analyze()

Modified the following up to line 734:

* Apply consistent code style.
* Removed a few commented items.

The plugin compiles and runs.

* Fourth pass at VHbbAnalysis::Analyze()

* The method now conforms to a consistent code style.
* All tabs have been replaced with spaces.

The plugin compiles and runs. Ready to begin adding in the 0-lepton
control sample.

* Begin adding 0-lepton TTbar Control Sample

Modified the following files:

* plugins/VHbbAnalysis::Analyze()
  This houses the main logic for defining control samples. So far, only
  a boolean local variable named base0LepCSSelection has been defined.
* VHbbAnalysis/cfg/earlybranches.txt
  Removed spurious whitespace at the end of lines.
* VHbbAnalysis/cfg/existingbranches.txt
  Removed spurious whitespace at the end of lines and added the
  following branches to the list:
  - HCMVAV2_mass
  - HCMVAV2_pt
  - Jet_btagCMVAV2
  - Flag_goodVertices
  - Flag_GlobalTightHalo2016Filter
  - Flag_HBHENoiseFilter
  - Flag_HBHENoiseIsoFilter
  - Flag_EcalDeadCellTriggerPrimitiveFilter

The plugin compiles and runs. Note that the cuts are hard-coded for
this first attempt, this may be modified in the future.

* Continue adding 0-lepton TTbar Control Sample

Added the computation of intermediate quantities used to cut on the
events for 0-lepton control samples within VHbbAnalysis.cc.

Added the branch "V_phi" to the list in existingbranches.txt.

The plugin compiles and runs.

* Finish adding all 0-lepton Control Samples

Finished adding the logic defining the 0-lepton controls samples to the
VHbbAnalysis.cc plugin.

Added the following branches to existingbranches.txt:
* HCMVAV2_reg_phi
* HCMVAV2_reg_mass
* tkMet_phi

The plugin compiles and runs.

* adding desy samples

* adding desy samples (uncommented)

* resubmit option to check the input files

* adding skims at DESY

* hadd script for desy

* auto-detecting site...

* Exporting a default PATH in the condor runscript for DESY

* added missing settings (was producing seg faults)

* Add missing settings to settings_CS.txt

As suggested by @scooperstein, the segfault encountered while running
with the control sample settings file "settings_CS.txt" were in fact due
to new settings which had been declared in "settings.txt" and used in
the VHbbAnalysis::Analyze() method but were not similarly declared in
"settings_CS.txt".

The following settings were added:
* e1ptcut
* e2ptcut
* mu1ptcut
* mu2ptcut
* zmasslow
* zmasshigh

Running the following command with the control sample config file no
longer segfaults:

```bash
python RunSample.py vhbb_config.txt TT_powheg "" output.root
```

* Effectively remove nAddJetsCut in settings_CS.txt

* config for plotting the control region updated

* config for plotting the control region updated #2

* set nAddJetsCut to large value for CS

* updating settings_CS to latest definitions

* hadd all data files into one

* inverting MET cut for CS again

* whitespace

* adding multiprocessing support to plotFromTree.py with -N <n_cpus> option

* cleanup and code comments

* adding 2-lepton channel control regions, cuts also hard-coded for now

* removing atLeastOnePassesCMVA from preselection to include lf control regions

* print line for zombie check. Only way to find files that couldn't be opened

* print line for zombie check. Only way to find files that couldn't be opened

* using Vtype and isZll variables for 2-lepton CS

* resetting old weights

* latest skim

* desy specific configs

* exit if am cannot be initialise

* exit if am cannot be initialise

* latest skim, no-btag-in-preselection fixed

* removing leftover Vtype requirement from btag requirement

* no V_mass cut in control regions

* adding looser two-electron selection to preselection